Split server into multiple files
